requirements.txt
requests==2.26.0
openai==0.27.0
pyyaml==6.0.1
python-dotenv
Cython
anthropic
groq
google-generativeai